refactor blueprints (3/6): extract below_deck blueprint
Seven /below-deck/* routes move to blueprints/below_deck.py. The
4am ET auto-clear logic stays inline in below_deck() (the page
route) — it's only used there. No new helpers extracted; the
blueprint is straightforward CRUD on SQLite tasks where
project_id IS NULL.

Routes moved:
- GET /below-deck (page render with auto-clear)
- GET /below-deck/count
- POST /below-deck/add
- POST /below-deck/complete
- POST /below-deck/delete
- POST /below-deck/clear-completed
- POST /below-deck/reorder

`url_for('login')` retained — migrates in sub-phase 5 with cockpit.

Smoke-tested: /below-deck, /below-deck/count, /today/data, /publish
all 200 (last two are cross-blueprint sanity).
